requirements.txt
Flask==3.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
PyYAML==6.0.1
tqdm==4.66.1
matplotlib==3.7.3
seaborn==0.12.2
pandas==2.0.3
scipy==1.11.3
Werkzeug==3.0.1
gunicorn==21.2.0